Исправление замечаний №2 от "SnurnicynKonstantin"
