+++ src/main/java/school/faang/user_service/controller/recommendation/RecommendationRequestController.java
@@ -18,7 +18,7 @@ import java.util.List;
 
 @RestController
 @RequiredArgsConstructor
-@RequestMapping
+@RequestMapping("/api/recommendations")
 public class RecommendationRequestController {
 
     private final RecommendationRequestService recommendationRequestService;
