+++ src/main/java/school/faang/user_service/filter/recommendation/MessagePatternFilter.java
@@ -1,19 +1,18 @@
-package school.faang.user_service.controller.filter;
-
+package school.faang.user_service.filter.recommendation;
 
 import school.faang.user_service.dto.RequestFilterDto;
 import school.faang.user_service.entity.recommendation.RecommendationRequest;
 
 import java.util.stream.Stream;
 
-public class ReceiverIdFilter implements RecommendationFilter {
+public class MessagePatternFilter implements RecommendationFilter {
     @Override
     public boolean isApplicable(RequestFilterDto filters) {
-        return filters.receiverId() != null;
+        return filters.messagePattern() != null && !filters.messagePattern().isBlank();
     }
 
     @Override
     public Stream<RecommendationRequest> apply(Stream<RecommendationRequest> requests, RequestFilterDto filters) {
-        return requests.filter(request -> request.getReceiver().getId().equals(filters.receiverId()));
+        return requests.filter(request -> request.getMessage().contains(filters.messagePattern()));
     }
 }
