+++ src/main/java/school/faang/user_service/filter/recommendation/ReceiverIdFilter.java
@@ -1,18 +1,19 @@
-package school.faang.user_service.controller.filter;
+package school.faang.user_service.filter.recommendation;
+
 
 import school.faang.user_service.dto.RequestFilterDto;
 import school.faang.user_service.entity.recommendation.RecommendationRequest;
 
 import java.util.stream.Stream;
 
-public class MessagePatternFilter implements RecommendationFilter {
+public class ReceiverIdFilter implements RecommendationFilter {
     @Override
     public boolean isApplicable(RequestFilterDto filters) {
-        return filters.messagePattern() != null && !filters.messagePattern().isBlank();
+        return filters.receiverId() != null;
     }
 
     @Override
     public Stream<RecommendationRequest> apply(Stream<RecommendationRequest> requests, RequestFilterDto filters) {
-        return requests.filter(request -> request.getMessage().contains(filters.messagePattern()));
+        return requests.filter(request -> request.getReceiver().getId().equals(filters.receiverId()));
     }
 }
