+++ src/main/java/school/faang/user_service/filter/recommendation/RequesterIdFilter.java
@@ -1,4 +1,4 @@
-package school.faang.user_service.controller.filter;
+package school.faang.user_service.filter.recommendation;
 
 import school.faang.user_service.dto.RequestFilterDto;
 import school.faang.user_service.entity.recommendation.RecommendationRequest;
