+++ src/main/java/school/faang/user_service/mapper/RecommendationMapper.java
@@ -1,6 +1,7 @@
 package school.faang.user_service.mapper;
 
 import org.mapstruct.Mapper;
+import org.mapstruct.Mapping;
 import school.faang.user_service.dto.RecommendationRequestDto;
 import school.faang.user_service.dto.RecommendationResponseDto;
 import school.faang.user_service.entity.Skill;
@@ -10,8 +11,16 @@ import school.faang.user_service.entity.recommendation.SkillRequest;
 @Mapper(componentModel = "spring")
 public interface RecommendationMapper {
 
+    @Mapping(target = "skills", ignore = true)
+    @Mapping(target = "createdAt", ignore = true)
+    @Mapping(target = "id", ignore = true)
+    @Mapping(target = "updatedAt", ignore = true)
+    @Mapping(target = "rejectionReason", ignore = true)
+    @Mapping(target = "status", ignore = true)
     RecommendationRequest toEntity(RecommendationRequestDto dto);
 
+    @Mapping(target = "requesterId", source = "requester.id")
+    @Mapping(target = "receiverId", source = "receiver.id")
     RecommendationResponseDto toDto(RecommendationRequest entity);
 
     default SkillRequest map(String skillName) {
